[frontend] History tab top margin / spacing  (#9568)
